WIP status component improvements
