+++ plugins/ros/src/components/riScInfo/MigrationDialog.tsx
@@ -1,0 +1,85 @@
+import React, { ReactComponentElement, useState } from 'react';
+import Alert from '@mui/material/Alert';
+import Button from '@mui/material/Button';
+import Dialog from '@mui/material/Dialog';
+import DialogContent from '@mui/material/DialogContent';
+import DialogTitle from '@mui/material/DialogTitle';
+import DialogActions from '@mui/material/DialogActions';
+import Link from '@mui/material/Link';
+import Typography from '@mui/material/Typography';
+import Checkbox from '@mui/material/Checkbox';
+import { useTranslationRef } from '@backstage/core-plugin-api/alpha';
+import FormControlLabel from '@mui/material/FormControlLabel';
+import Box from '@mui/material/Box';
+import { pluginRiScTranslationRef } from '../../utils/translations';
+import { MigrationVersions } from '../../utils/types';
+import { dialogActions } from '../common/mixins';
+
+interface RiScMigrationDialogProps {
+  openDialog: boolean;
+  handleCancel: () => void;
+  handleUpdate: () => void;
+  migrationVersions?: MigrationVersions;
+}
+
+export const RiScMigrationDialog = ({
+  openDialog,
+  handleCancel,
+  handleUpdate,
+  migrationVersions,
+}: RiScMigrationDialogProps): ReactComponentElement<any> => {
+  const { t } = useTranslationRef(pluginRiScTranslationRef);
+
+  const [saveMigration, setSaveMigration] = useState<boolean>(false);
+
+  const handleCheckboxInput = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setSaveMigration(event.target.checked);
+  };
+
+  return (
+    <Dialog open={openDialog}>
+      <DialogTitle>{t('migrationDialog.title')}</DialogTitle>
+      <DialogContent>
+        <Box sx={{ marginBottom: '16px' }}>
+          <Typography>
+            {t('migrationDialog.description')}
+            {migrationVersions?.toVersion} {t('migrationDialog.description2')}{' '}
+            {migrationVersions?.fromVersion}
+            {t('migrationDialog.description3')}
+            <Link
+              underline="always"
+              target="_blank"
+              href="https://github.com/kartverket/backstage-plugin-risk-scorecard-backend/blob/main/docs/schemaChangelog.md"
+            >
+              {t('migrationDialog.changelog')}
+            </Link>{' '}
+            {t('migrationDialog.description4')}
+          </Typography>
+        </Box>
+        <Alert severity="info" icon={false}>
+          <FormControlLabel
+            control={
+              <Checkbox
+                checked={saveMigration}
+                onChange={handleCheckboxInput}
+              />
+            }
+            label={t('migrationDialog.checkboxLabel')}
+          />
+        </Alert>
+      </DialogContent>
+      <DialogActions sx={dialogActions}>
+        <Button
+          variant="contained"
+          onClick={handleUpdate}
+          disabled={!saveMigration}
+        >
+          {t('dictionary.confirm')}
+        </Button>
+        <Button variant="outlined" color="primary" onClick={handleCancel}>
+          {t('dictionary.cancel')}
+        </Button>
+      </DialogActions>
+    </Dialog>
+  );
+};
