+++ plugins/ros/src/components/riScInfo/riScStatus/RiScStatusComponent.tsx
@@ -1,24 +1,15 @@
-import React, { ReactComponentElement, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import {
   DifferenceFetchState,
-  MigrationVersions,
   RiScStatus,
   RiScWithMetadata,
 } from '../../../utils/types';
-import Alert from '@mui/material/Alert';
 import Button from '@mui/material/Button';
-import Dialog from '@mui/material/Dialog';
-import DialogContent from '@mui/material/DialogContent';
-import DialogTitle from '@mui/material/DialogTitle';
-import DialogActions from '@mui/material/DialogActions';
 import Link from '@mui/material/Link';
 import Typography from '@mui/material/Typography';
-import Checkbox from '@mui/material/Checkbox';
 import { useTranslationRef } from '@backstage/core-plugin-api/alpha';
 import { pluginRiScTranslationRef } from '../../../utils/translations';
 import CheckIcon from '@mui/icons-material/Check';
-import FormControlLabel from '@mui/material/FormControlLabel';
-import { dialogActions } from '../../common/mixins';
 import { InfoCard } from '@backstage/core-components';
 import { PullRequestSvg } from '../../common/Icons';
 import { useRiScs } from '../../../contexts/RiScContext';
@@ -26,182 +17,9 @@ import { subtitle1 } from '../../common/typography';
 import Box from '@mui/material/Box';
 import { WarningAmberOutlined } from '@mui/icons-material';
 import { useAuthenticatedFetch } from '../../../utils/hooks';
-import { RiScDifferenceDialog } from './RiScDifferenceDialog';
-
-interface RiScPublishDialogProps {
-  openDialog: boolean;
-  handleCancel: () => void;
-  handlePublish: () => void;
-  differenceFetchState: DifferenceFetchState;
-}
-
-const RiScPublishDialog = ({
-  openDialog,
-  handleCancel,
-  handlePublish,
-  differenceFetchState,
-}: RiScPublishDialogProps): ReactComponentElement<any> => {
-  const { t } = useTranslationRef(pluginRiScTranslationRef);
-
-  const [riskOwnerApproves, setRiskOwnerApproves] = useState<boolean>(false);
-
-  const handleCheckboxInput = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setRiskOwnerApproves(event.target.checked);
-  };
-
-  return (
-    <Dialog open={openDialog}>
-      <DialogTitle>{t('publishDialog.title')}</DialogTitle>
-      <DialogContent>
-        <>
-          <RiScDifferenceDialog differenceFetchState={differenceFetchState} />
-          <Alert severity="info" icon={false}>
-            <FormControlLabel
-              control={
-                <Checkbox
-                  color="primary"
-                  checked={riskOwnerApproves}
-                  onChange={handleCheckboxInput}
-                />
-              }
-              label={t('publishDialog.checkboxLabel')}
-            />
-          </Alert>
-        </>
-      </DialogContent>
-      <DialogActions sx={dialogActions}>
-        <Button
-          variant="contained"
-          color="primary"
-          onClick={handlePublish}
-          disabled={!riskOwnerApproves}
-        >
-          {t('dictionary.confirm')}
-        </Button>
-        <Button variant="outlined" color="primary" onClick={handleCancel}>
-          {t('dictionary.cancel')}
-        </Button>
-      </DialogActions>
-    </Dialog>
-  );
-};
-
-interface RiScMigrationDialogProps {
-  openDialog: boolean;
-  handleCancel: () => void;
-  handleUpdate: () => void;
-  migrationVersions?: MigrationVersions;
-}
-
-const RiScMigrationDialog = ({
-  openDialog,
-  handleCancel,
-  handleUpdate,
-  migrationVersions,
-}: RiScMigrationDialogProps): ReactComponentElement<any> => {
-  const { t } = useTranslationRef(pluginRiScTranslationRef);
-
-  const [saveMigration, setSaveMigration] = useState<boolean>(false);
-
-  const handleCheckboxInput = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setSaveMigration(event.target.checked);
-  };
-
-  return (
-    <Dialog open={openDialog}>
-      <DialogTitle>{t('migrationDialog.title')}</DialogTitle>
-      <DialogContent>
-        <Box sx={{ marginBottom: '16px' }}>
-          <Typography>
-            {t('migrationDialog.description')}
-            {migrationVersions?.toVersion} {t('migrationDialog.description2')}{' '}
-            {migrationVersions?.fromVersion}
-            {t('migrationDialog.description3')}
-            <Link
-              underline="always"
-              target="_blank"
-              href="https://github.com/kartverket/backstage-plugin-risk-scorecard-backend/blob/main/docs/schemaChangelog.md"
-            >
-              {t('migrationDialog.changelog')}
-            </Link>{' '}
-            {t('migrationDialog.description4')}
-          </Typography>
-        </Box>
-        <Alert severity="info" icon={false}>
-          <FormControlLabel
-            control={
-              <Checkbox
-                checked={saveMigration}
-                onChange={handleCheckboxInput}
-              />
-            }
-            label={t('migrationDialog.checkboxLabel')}
-          />
-        </Alert>
-      </DialogContent>
-      <DialogActions sx={dialogActions}>
-        <Button
-          variant="contained"
-          onClick={handleUpdate}
-          disabled={!saveMigration}
-        >
-          {t('dictionary.confirm')}
-        </Button>
-        <Button variant="outlined" color="primary" onClick={handleCancel}>
-          {t('dictionary.cancel')}
-        </Button>
-      </DialogActions>
-    </Dialog>
-  );
-};
-
-const RosAcceptance = ({
-  status,
-  migration,
-}: {
-  status: RiScStatus;
-  migration?: boolean;
-}): React.JSX.Element => {
-  const { t } = useTranslationRef(pluginRiScTranslationRef);
-  if (migration) {
-    return (
-      <>
-        <Typography paragraph sx={subtitle1}>
-          <WarningAmberOutlined
-            fontSize="medium"
-            sx={{ transform: 'translateY(5px)', marginTop: '5px' }}
-          />{' '}
-          {t('rosStatus.statusBadge.migration.title')}
-        </Typography>
-        <Typography>
-          {t('rosStatus.statusBadge.migration.description')}
-        </Typography>
-      </>
-    );
-  }
-  switch (status) {
-    case RiScStatus.Draft:
-      return (
-        <Typography paragraph sx={subtitle1}>
-          {t('rosStatus.statusBadge.missing')}
-        </Typography>
-      );
-    case RiScStatus.SentForApproval:
-    case RiScStatus.Published:
-      return (
-        <Typography paragraph sx={subtitle1}>
-          <CheckIcon fontSize="medium" sx={{ transform: 'translateY(5px)' }} />{' '}
-          {t('rosStatus.statusBadge.approved')}
-        </Typography>
-      );
-    default:
-      return (
-        <Typography paragraph sx={subtitle1}>
-          {t('rosStatus.statusBadge.error')}
-        </Typography>
-      );
-  }
-};
+import Progress from './Progress';
+import { RiScMigrationDialog } from '../MigrationDialog';
+import { RiScPublishDialog } from '../PublishDialog';
 
 const emptyDifferenceFetchState: DifferenceFetchState = {
   differenceState: {
@@ -289,16 +107,91 @@ export const RiScStatusComponent = ({
     setDifferenceFetchState(emptyDifferenceFetchState);
   }, [selectedRiSc]);
 
+  const [status, setStatus] = useState<1 | 2 | 3>(1);
+
+  useEffect(() => {
+    if (selectedRiSc.status === RiScStatus.Draft) {
+      setStatus(1);
+    } else if (selectedRiSc.status === RiScStatus.SentForApproval) {
+      setStatus(2);
+    } else if (selectedRiSc.status === RiScStatus.Published) {
+      setStatus(3);
+    }
+  }, [selectedRiSc.status]);
+
+  const migration = selectedRiSc.migrationStatus?.migrationChanges;
+  // const migration = true;
+
   return (
     <InfoCard>
       <Typography variant="h5">Status</Typography>
 
-      <RosAcceptance
-        status={selectedRiSc.status}
-        migration={selectedRiSc.migrationStatus?.migrationChanges}
-      />
-      {selectedRiSc.status === RiScStatus.SentForApproval &&
-        !selectedRiSc.migrationStatus?.migrationChanges && (
+      <Box
+        display="flex"
+        flexDirection="row"
+        alignItems="center"
+        gap={2}
+        justifyContent="space-between"
+        m={1}
+      >
+        <Progress step={status} />
+
+        {/* Migration */}
+
+        {migration && (
+          <>
+            <Typography paragraph sx={subtitle1}>
+              <WarningAmberOutlined
+                fontSize="medium"
+                sx={{ transform: 'translateY(5px)', marginTop: '5px' }}
+              />{' '}
+              {t('rosStatus.statusBadge.migration.title')}
+            </Typography>
+            <Typography>
+              {t('rosStatus.statusBadge.migration.description')}
+            </Typography>
+            <Button
+              color="primary"
+              variant="contained"
+              onClick={() => setMigrationDialogIsOpen(!migrationDialogIsOpen)}
+              sx={{ display: 'block', marginLeft: 'auto' }}
+            >
+              <Typography variant="button">
+                {t('rosStatus.moreInformationButton')}
+              </Typography>
+            </Button>
+            <RiScMigrationDialog
+              openDialog={migrationDialogIsOpen}
+              handleUpdate={handleUpdate}
+              handleCancel={() => setMigrationDialogIsOpen(false)}
+              migrationVersions={
+                selectedRiSc.migrationStatus?.migrationVersions
+              }
+            />
+          </>
+        )}
+
+        {/* Draft (1) */}
+
+        {!migration && status === 1 && (
+          <>
+            <Typography paragraph sx={subtitle1}>
+              {t('rosStatus.statusBadge.missing')}
+            </Typography>
+            <Button
+              color="primary"
+              variant="contained"
+              onClick={handleOpenPublishRiScDialog}
+              sx={{ display: 'block', marginLeft: 'auto' }}
+            >
+              {t('rosStatus.approveButton')}
+            </Button>
+          </>
+        )}
+
+        {/* SentForApproval (2) */}
+
+        {!migration && status === 2 && (
           <Typography sx={{ fontWeight: 700 }} paragraph variant="subtitle1">
             <PullRequestSvg />
             {t('rosStatus.prStatus')}
@@ -307,37 +200,28 @@ export const RiScStatusComponent = ({
             </Link>
           </Typography>
         )}
-      {selectedRiSc.status === RiScStatus.Draft &&
-        !selectedRiSc.migrationStatus?.migrationChanges && (
-          <Button
-            color="primary"
-            variant="contained"
-            onClick={handleOpenPublishRiScDialog}
-            sx={{ display: 'block', marginLeft: 'auto' }}
-          >
-            {t('rosStatus.approveButton')}
-          </Button>
+
+        {/* Published (3) */}
+
+        {!migration && status === 3 && (
+          <Typography paragraph sx={subtitle1}>
+            <CheckIcon
+              fontSize="medium"
+              sx={{ transform: 'translateY(5px)' }}
+            />{' '}
+            {t('rosStatus.statusBadge.approved')}
+          </Typography>
         )}
-      {selectedRiSc.migrationStatus?.migrationChanges && (
-        <>
-          <Button
-            color="primary"
-            variant="contained"
-            onClick={() => setMigrationDialogIsOpen(!migrationDialogIsOpen)}
-            sx={{ display: 'block', marginLeft: 'auto' }}
-          >
-            <Typography variant="button">
-              {t('rosStatus.moreInformationButton')}
-            </Typography>
-          </Button>
-          <RiScMigrationDialog
-            openDialog={migrationDialogIsOpen}
-            handleUpdate={handleUpdate}
-            handleCancel={() => setMigrationDialogIsOpen(false)}
-            migrationVersions={selectedRiSc.migrationStatus?.migrationVersions}
-          />
-        </>
+      </Box>
+
+      {/* Error */}
+
+      {!selectedRiSc && (
+        <Typography paragraph sx={subtitle1}>
+          {t('rosStatus.statusBadge.error')}
+        </Typography>
       )}
+
       <RiScPublishDialog
         openDialog={publishRiScDialogIsOpen}
         handlePublish={handleApproveAndPublish}
